requirements.txt
autopep8==2.0.0
pycodestyle==2.9.1
python-lorem==1.2.0
tomli==2.0.1
uuid==1.30
